refactor: 유저 정보 페이지 비즈니스 로직 분리
